components/rdb-exporter/requirements.txt
asyncio==3.4.3
awsiotsdk==1.12.2
cerberus==1.3.4